Refactored according to directory changes and added improved json reads.
