+++ app/v1/models/response/common/CalculationDetail.scala
@@ -14,30 +14,12 @@
  * limitations under the License.
  */
 
-package v1.models.des.selfAssessment.componentObjects
+package v1.models.response.common
 
-import play.api.libs.json.{JsSuccess, JsValue, Json}
-import support.UnitSpec
+import play.api.libs.json.{Json, OFormat}
 
-class CalculationSummarySpec extends UnitSpec {
+case class CalculationDetail(incomeTax: String)
 
-  val json: JsValue = Json.parse(
-    """
-      |{
-      | "incomeTax" : ""
-      |}
-    """.stripMargin)
-
-  val model = CalculationSummary("")
-
-  "CalculationDetail" should {
-
-    "write to json correctly" in {
-      Json.toJson(model) shouldBe json
-    }
-
-    "read from json correctly" in {
-      json.validate[CalculationSummary] shouldBe JsSuccess(model)
-    }
-  }
+object CalculationDetail {
+  implicit val format: OFormat[CalculationDetail] = Json.format[CalculationDetail]
 }
